Add output diagnostic for resetting ocean temperatures below sea-ice threshold (-1.8C).  This is now singout(9).  Relates to ticket #1.
Added test for isothermal layers in new subroutine CHECK_ISOTHERMAL (#2).
Test is based on integrated absolute value of difference in temperature
between surface and level number specified by isotherm_bottom in
NAME_FORCING.  If the integrated absolute value is less than the
specified threshold isotherm_threshold (also in NAME_FORCING), the
T and S profiles are reset to climatology and a flag is set to the value
of the integrated absolute value; this flag is available as singout(10).
Use L_NO_ISOTHERM in NAME_FORCING to enable.  Enabling this requires
specifying a filename for 3D ocean temperature (ocnT_file) and salinity
(sal_file) files.